feat: react router lazy route splitting (#970)
