+++ src/components/app/App.jsx
@@ -1,22 +1,5 @@
-import { lazy, useEffect } from 'react';
-import {
-  createBrowserRouter,
-  createRoutesFromElements,
-  RouterProvider,
-  Route,
-  Outlet,
-  useRouteError,
-  generatePath,
-  useAsyncError,
-  Link,
-} from 'react-router-dom';
-import { Container } from '@edx/paragon';
-import {
-  AppProvider,
-  AuthenticatedPageRoute,
-  ErrorPage,
-} from '@edx/frontend-platform/react';
-import { logError } from '@edx/frontend-platform/logging';
+import { RouterProvider } from 'react-router-dom';
+import { AppProvider } from '@edx/frontend-platform/react';
 import {
   QueryCache,
   QueryClient,
@@ -24,31 +7,17 @@ import {
 } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
-import { queryCacheOnErrorHandler, defaultQueryClientRetryHandler } from '../../utils/common';
-import extractNamedExport from '../../utils/extract-named-export';
 import {
-  makeDashboardLoader,
-  makeUpdateActiveEnterpriseCustomerUserLoader,
-  makeCourseLoader,
-  makeRootLoader,
-} from './routes/loaders';
-import {
-  useEnterpriseLearner,
-  useCourseRedemptionEligibility,
-  useEnterpriseCustomerUserSubsidies,
-  useCourseMetadata,
-  useEnterpriseCourseEnrollments,
-  useUserEntitlements,
-} from './data';
-import Root from './Root';
-import Layout from './Layout';
+  queryCacheOnErrorHandler,
+  defaultQueryClientRetryHandler,
+} from '../../utils/common';
+// import extractNamedExport from '../../utils/extract-named-export';
 
-/* eslint-disable no-unused-vars */
-const EnterpriseCustomerRedirect = lazy(() => import(/* webpackChunkName: "enterprise-customer-redirect" */ '../enterprise-redirects/EnterpriseCustomerRedirect'));
-const EnterprisePageRedirect = lazy(() => import(/* webpackChunkName: "enterprise-page-redirect" */ '../enterprise-redirects/EnterprisePageRedirect'));
-const NotFoundPage = lazy(() => import(/* webpackChunkName: "not-found" */ '../NotFoundPage'));
-const EnterpriseAppPageRoutes = lazy(() => import(/* webpackChunkName: "enterprise-app-routes" */ './EnterpriseAppPageRoutes'));
-const EnterpriseInvitePage = lazy(() => extractNamedExport(import(/* webpackChunkName: "enterprise-invite" */ '../enterprise-invite'), 'EnterpriseInvitePage'));
+import createAppRouter from './data/createAppRouter';
+
+/* eslint-disable max-len */
+// const EnterpriseAppPageRoutes = lazy(() => import(/* webpackChunkName: "enterprise-app-routes" */ './EnterpriseAppPageRoutes'));
+// const EnterpriseInvitePage = lazy(() => extractNamedExport(import(/* webpackChunkName: "enterprise-invite" */ '../enterprise-invite'), 'EnterpriseInvitePage'));
 /* eslint-enable no-unused-vars */
 
 // Create a query client for @tanstack/react-query
@@ -69,175 +38,7 @@ const queryClient = new QueryClient({
   },
 });
 
-const Dashboard = () => {
-  const { data: enterpriseCustomerUserSubsidies } = useEnterpriseCustomerUserSubsidies();
-  const {
-    data: enterpriseCourseEnrollments,
-    isLoading: isLoadingEnterpriseCourseEnrollments,
-    isFetching: isFetchingEnterpriseCourseEnrollments,
-  } = useEnterpriseCourseEnrollments();
-
-  return (
-    <Container size="lg" className="py-4">
-      <h2>Dashboard</h2>
-      <pre>
-        {JSON.stringify(
-          {
-            enterpriseCourseEnrollments: {
-              isLoading: isLoadingEnterpriseCourseEnrollments,
-              isFetching: isFetchingEnterpriseCourseEnrollments,
-              count: enterpriseCourseEnrollments?.length,
-            },
-            enterpriseCustomerUserSubsidies,
-          },
-          null,
-          2,
-        )}
-      </pre>
-    </Container>
-  );
-};
-
-const Search = () => {
-  const { data: enterpriseCustomerUserSubsidies } = useEnterpriseCustomerUserSubsidies();
-  const { data: { enterpriseCustomer } } = useEnterpriseLearner();
-  return (
-    <Container size="lg" className="py-4">
-      <h2>Search</h2>
-      <Link
-        to={generatePath('/:enterpriseSlug/course/:courseKey', {
-          enterpriseSlug: enterpriseCustomer.slug,
-          courseKey: 'edX+DemoX',
-        })}
-      >
-        Course
-      </Link>
-      <br />
-      <br />
-      <pre>{JSON.stringify(enterpriseCustomerUserSubsidies, null, 2)}</pre>
-    </Container>
-  );
-};
-
-const Course = () => {
-  const { data: courseMetadata } = useCourseMetadata();
-  const { data: courseRedemptionEligiblity } = useCourseRedemptionEligibility();
-  const {
-    data: enterpriseCourseEnrollments,
-    isLoading: isLoadingEnterpriseCourseEnrollments,
-    isFetching: isFetchingEnterpriseCourseEnrollments,
-  } = useEnterpriseCourseEnrollments();
-
-  const {
-    data: userEntitlements,
-    isLoading: isUserEntitlementsLoading,
-    isFetching: isUserEntitlementsFetching,
-  } = useUserEntitlements();
-
-  if (!courseMetadata) {
-    return <NotFoundPage />;
-  }
-
-  return (
-    <Container size="lg" className="py-4">
-      <h2>Course</h2>
-      <pre>
-        {JSON.stringify(
-          {
-            courseMetadata: {
-              title: courseMetadata.title,
-              enrollmentUrl: courseMetadata.enrollmentUrl,
-            },
-            redeemableSubsidyAccessPolicy: courseRedemptionEligiblity.find(
-              ({ canRedeem }) => canRedeem,
-            )?.redeemableSubsidyAccessPolicy?.uuid ?? null,
-            enterpriseCourseEnrollments: {
-              isLoading: isLoadingEnterpriseCourseEnrollments,
-              isFetching: isFetchingEnterpriseCourseEnrollments,
-              count: enterpriseCourseEnrollments?.length || 0,
-            },
-            userEntitlements: {
-              isLoading: isUserEntitlementsLoading,
-              isFetching: isUserEntitlementsFetching,
-              count: userEntitlements?.length || 0,
-            },
-          },
-          null,
-          2,
-        )}
-      </pre>
-    </Container>
-  );
-};
-
-const retrieveErrorBoundaryErrorMessage = (error) => {
-  if (!error) {
-    return null;
-  }
-  if (error.customAttributes) {
-    return error.customAttributes.httpErrorResponseData;
-  }
-  return error.message;
-};
-
-const RouteErrorBoundary = () => {
-  const routeError = useRouteError();
-  const asyncError = useAsyncError();
-
-  useEffect(() => {
-    if (routeError) {
-      logError(routeError);
-    }
-  }, [routeError]);
-
-  useEffect(() => {
-    if (asyncError) {
-      logError(asyncError);
-    }
-  }, [asyncError]);
-
-  const errorMessage = retrieveErrorBoundaryErrorMessage(routeError || asyncError);
-  return <ErrorPage message={errorMessage} />;
-};
-
-const router = createBrowserRouter(
-  createRoutesFromElements(
-    <Route
-      path="/"
-      element={<AuthenticatedPageRoute><Root /></AuthenticatedPageRoute>}
-      errorElement={<RouteErrorBoundary />}
-    >
-      <Route
-        path="/:enterpriseSlug?"
-        loader={makeUpdateActiveEnterpriseCustomerUserLoader(queryClient)}
-        element={<Outlet />}
-      >
-        <Route
-          path=""
-          loader={makeRootLoader(queryClient)}
-          element={<Layout />}
-        >
-          <Route
-            index
-            element={<Dashboard />}
-            loader={makeDashboardLoader(queryClient)}
-          />
-          <Route
-            path="search"
-            element={<Search />}
-          />
-          <Route
-            path=":courseType?/course/:courseKey/*"
-            element={<Course />}
-            loader={makeCourseLoader(queryClient)}
-          />
-          <Route path="*" element={<NotFoundPage />} />
-        </Route>
-      </Route>
-      <Route path="*" element={<NotFoundPage />} />
-    </Route>,
-  ),
-);
+const router = createAppRouter(queryClient);
 
 const App = () => (
   <QueryClientProvider client={queryClient}>
@@ -250,11 +51,8 @@ const App = () => (
       )}
       >
         <Routes>
-          <Route path="/" element={<AuthenticatedPageRoute><EnterpriseCustomerRedirect /></AuthenticatedPageRoute>} />
-          <Route path="/r/*" element={<AuthenticatedPageRoute><EnterprisePageRedirect /></AuthenticatedPageRoute>} />
           <Route path="/invite/:enterpriseCustomerInviteKey" element={<PageWrap><EnterpriseInvitePage /></PageWrap>} />
           <Route path="/:enterpriseSlug/*" element={<EnterpriseAppPageRoutes />} />
-          <Route path="*" element={<PageWrap><NotFoundPage /></PageWrap>} />
         </Routes>
       </Suspense> */}
     </AppProvider>
