+++ src/components/app/routes/loaders/courseLoader.js
@@ -3,39 +3,9 @@ import {
   makeCanRedeemQuery,
   makeCourseMetadataQuery,
   makeEnterpriseCourseEnrollmentsQuery,
-  makeEnterpriseLearnerQuery,
   makeUserEntitlementsQuery,
 } from '../queries';
-
-// TODO: abstract out somewhere more common
-export async function extractActiveEnterpriseId({
-  queryClient,
-  authenticatedUser,
-  enterpriseSlug,
-}) {
-  // Retrieve linked enterprise customers for the current user from query cache
-  // or fetch from the server if not available.
-  const linkedEnterpriseCustomersQuery = makeEnterpriseLearnerQuery(authenticatedUser.username, enterpriseSlug);
-  const enterpriseLearnerData = await queryClient.ensureQueryData(linkedEnterpriseCustomersQuery);
-  const {
-    activeEnterpriseCustomer,
-    allLinkedEnterpriseCustomerUsers,
-  } = enterpriseLearnerData;
-
-  if (!enterpriseSlug) {
-    return activeEnterpriseCustomer.uuid;
-  }
-
-  const foundEnterpriseIdForSlug = allLinkedEnterpriseCustomerUsers.find(
-    (enterpriseCustomerUser) => enterpriseCustomerUser.enterpriseCustomer.slug === enterpriseSlug,
-  )?.enterpriseCustomer.uuid;
-
-  if (foundEnterpriseIdForSlug) {
-    return foundEnterpriseIdForSlug;
-  }
-
-  throw new Error(`Could not find enterprise customer for user ${authenticatedUser.userId} and slug ${enterpriseSlug}`);
-}
+import extractEnterpriseId from './extractEnterpriseId';
 
 /**
  * TODO
@@ -47,7 +17,7 @@ export default function makeCourseLoader(queryClient) {
     const authenticatedUser = await ensureAuthenticatedUser();
     const { courseKey, enterpriseSlug } = params;
 
-    const enterpriseId = await extractActiveEnterpriseId({
+    const enterpriseId = await extractEnterpriseId({
       queryClient,
       authenticatedUser,
       enterpriseSlug,
