--- src/components/enterprise-redirects/index.js
@@ -1,2 +1,0 @@
-export { default as EnterpriseCustomerRedirect } from './EnterpriseCustomerRedirect';
-export { default as EnterprisePageRedirect } from './EnterprisePageRedirect';
